Step3の3まで動作するように修正
